Base off alpine, not ruby
So .... the ruby base image is really heavy. This is because it's based
on Ubuntu?

Rebuilding image off alpine, a really tiny distro.
Also reorganized Dockerfile based on best practices, which I think
is structured to minimize the number of layer regenerations. Ideally
now I'd only regenerate a layer when I change my corpus model.